+++ autosubmit/job/job.py
@@ -49,7 +49,6 @@ if TYPE_CHECKING:
 
 Log.get_logger("Autosubmit")
 
-
 # A wrapper for encapsulate threads , TODO: Python 3+ to be replaced by the < from concurrent.futures >